+++ Program.cs
@@ -11,7 +11,7 @@ namespace JALOKA
             // To customize application configuration such as set high DPI settings or default font,
             // see https://aka.ms/applicationconfiguration.
             ApplicationConfiguration.Initialize();
-            Application.Run(new FormLogin());
+            Application.Run(new V_register());
         }
     }
 }
